move JWT to prejoin closes #83
